Move addr.rs and rom_slice.rs into snes_utils module
