+++ smwe_project/src/lib.rs
@@ -1,10 +1,10 @@
 use std::{cell::RefCell, rc::Rc};
 
-use smwe_rom::Rom;
+use smwe_rom::SmwRom;
 
 pub struct Project {
     pub title:    String,
-    pub rom_data: Rom,
+    pub rom_data: SmwRom,
 }
 
 pub type ProjectRef = Rc<RefCell<Project>>;
