+++ smwe_rom/tests/with_rom_env.rs
@@ -1,11 +1,11 @@
 use std::env;
 
-use smwe_rom::Rom;
+use smwe_rom::SmwRom;
 
 #[test]
 #[ignore]
 fn test_with_rom_env() {
     let rom_path = env::var_os("ROM_PATH").expect("ROM_PATH not set");
     assert!(std::fs::metadata(&rom_path).expect("ROM_PATH invalid").is_file());
-    Rom::from_file(rom_path).expect("Rom parse error encountered");
+    SmwRom::from_file(rom_path).expect("Rom parse error encountered");
 }
